notes, clean up, move custom hook to utils
